make email a component so that it can be used in future homepage popup.
